move FabricAddon to upper level
